Clean architecture users refactor
