+++ src/backend/apps/FlowJudge.API.Service/Auth/Legal/CachedLegalLockManager.cs
@@ -1,7 +1,7 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 using FlowJudge.Common.Cache;
+using FlowJudge.Users.Application.Abstractions.Queries;
 using FlowJudge.Users.Application.Models;
-using FlowJudge.Users.Application.Queries;
 using System.Text;
 using System.Text.Json;
