+++ src/backend/apps/FlowJudge.API.Service/Program.cs
@@ -6,6 +6,7 @@ using FlowJudge.Common.Sql;
 using FlowJudge.Common.Sql.Migrations;
 using FlowJudge.Common.Utils;
 using FlowJudge.Users.Application;
+using FlowJudge.Users.Infrastructure;
 using FlowJudge.Workspaces.Application;
 using FlowJudge.Workspaces.Infrastructure;
 
@@ -63,7 +64,8 @@ builder.Services.AddCache(cfg =>
 
 builder.Services.AddTimeService();
 
-builder.Services.AddUsersModule();
+builder.Services.AddUsersApplication();
+builder.Services.AddUsersInfrastructure();
 
 builder.Services.AddWorkspacesApplication();
 builder.Services.AddWorkspacesInfrastructure();
