+++ src/backend/libs/modules/FlowJudge.Users.Application.Abstractions/Commands/AcceptPrivacyPolicyCommand.cs
@@ -1,6 +1,6 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 
-namespace FlowJudge.Users.Application.Commands
+namespace FlowJudge.Users.Application.Abstractions.Commands
 {
     public sealed record AcceptPrivacyPolicyCommand(Guid UserIdentityId, Guid PrivacyPolicyVersionId) : ICommand;
 }
