+++ src/backend/libs/modules/FlowJudge.Users.Application.Abstractions/Commands/AcceptTermsAndConditionsCommand.cs
@@ -1,6 +1,6 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 
-namespace FlowJudge.Users.Application.Commands
+namespace FlowJudge.Users.Application.Abstractions.Commands
 {
     public sealed record AcceptTermsAndConditionsCommand(Guid UserIdentityId, Guid TermsAndConditionsVersionId)
         : ICommand;
