+++ src/backend/libs/modules/FlowJudge.Users.Application.Abstractions/Models/UserLegalRequirements.cs
