+++ src/backend/libs/modules/FlowJudge.Users.Application.Abstractions/Queries/GetActualDocumentVersionQuery.cs
@@ -1,7 +1,7 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 using FlowJudge.Users.Application.Models;
 
-namespace FlowJudge.Users.Application.Queries
+namespace FlowJudge.Users.Application.Abstractions.Queries
 {
     public sealed record GetActualDocumentVersionQuery : IQuery<GetActualDocumentVersionQueryResult>
     {
