+++ src/backend/libs/modules/FlowJudge.Users.Application.Abstractions/Queries/GetUserLegalStateQuery.cs
@@ -1,6 +1,6 @@
 ﻿using FlowJudge.Common.Application.Mediator;
 
-namespace FlowJudge.Users.Application.Queries
+namespace FlowJudge.Users.Application.Abstractions.Queries
 {
     public sealed record GetUserLegalStateQuery(Guid UserIdentityId) : IQuery<GetUserLegalStateQueryResult>;
 }
