+++ src/backend/libs/modules/FlowJudge.Users.Infrastructure/Installer.cs
@@ -1,4 +1,5 @@
-﻿using FlowJudge.Users.Infrastructure.Repositories.DocumentVersions;
+﻿using FlowJudge.Users.Application.Abstractions.Ports;
+using FlowJudge.Users.Infrastructure.Repositories.DocumentVersions;
 using FlowJudge.Users.Infrastructure.Repositories.Users;
 using Microsoft.Extensions.DependencyInjection;
